Facebook login flow now works 1. Modified result_template.tpl to login user per facebook javascript sdk guidelines. Reasonably seamless flow to log into facebook (will do so automatically if logged in, which we should change), redirects  on successful login to autoregister as eMission user if not already in database (will want to add a check here to ensure on study list; I wasn't sure of how you guys do this so I just made something quick work). Right now, nothing else really happens. Probably wouldn't hurt to redirect directly to the homepage or choice screen.
