--- CFC_WebApp/clients/socialmedia/open_facebook/__init__.py
@@ -1,1 +1,0 @@
-__author__ = 'sidneyfeygin'
