+++ CFC_WebApp/tests/dao/TestFacebookUser.py
@@ -4,6 +4,7 @@ import os
 import json
 
 from facebook import GraphAPI, get_app_access_token
+from dao.socialmediauser import FacebookUser
 from dao.user import User
 
 sys.path.append("%s" % os.getcwd())
@@ -17,15 +18,17 @@ logging.basicConfig(level=logging.DEBUG)
 
 
 class TestFacebookUser(unittest.TestCase):
-    def testFromAccessToken(self):
-        from dao.SocialMediaUser import FacebookUser
+
+    def setUp(self):
         fb_config = open('clients/socialmedia/settings.json')
         fb_data = json.load(fb_config)['client_settings']
-        app_id = fb_data['fb_app_id']
-        secret_key = fb_data['fb_app_secret']
-        app_access_token = get_app_access_token(app_id, secret_key)
+        self.app_id = fb_data['fb_app_id']
+        self.secret_key = fb_data['fb_app_secret']
+
+    def testFromAccessToken(self):
+        app_access_token = get_app_access_token(self.app_id, self.secret_key)
         g = GraphAPI(app_access_token)
-        users = g.request('%s/accounts/test-users' % app_id)
+        users = g.request('%s/accounts/test-users' % self.app_id)
         user = users['data'][0]
         access_token = user['access_token']
         g2 = GraphAPI(access_token)
@@ -37,8 +40,10 @@ class TestFacebookUser(unittest.TestCase):
         self.assertEqual(resultPre, 1)
         self.assertEqual(resultReg, 0)
         user = User.register(email)
-        fbuser = FacebookUser.fromAccessToken(access_token)
+        fbuser = FacebookUser.from_access_token(access_token)
         self.assertEquals(fbuser.uuid, user.uuid)
+
+
 
 
 if __name__ == '__main__':
